requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langchain>=0.1.0
langchain-text-splitters>=0.3.0
langchain-groq>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
python-dotenv>=1.0.0
